# apps/python/openapi-spec.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "OAP LangGraph Runtime",
    "version": "0.1.0",
    "description": "\nRobyn-based LangGraph-compatible runtime for Open Agent Platform.\n\nThis API provides endpoints for managing assistants, threads, runs, and the store.\nIt implements the LangGraph API specification for compatibility with LangGraph clients.\n"
  },
...
  "paths": {
...
    "/a2a/{assistant_id}": {
      "post": {
        "tags": [
          "A2A"
        ],
        "summary": "A2A JSON-RPC",
        "description": "Communicate with an assistant using the Agent-to-Agent (A2A) Protocol over JSON-RPC 2.0.\nThis endpoint accepts a JSON-RPC envelope and dispatches based on `method`.\n\n**Supported Methods:**\n- `message/send`: Send a message and wait for the final Task result.\n- `message/stream`: Send a message and receive Server-Sent Events (SSE) JSON-RPC responses.\n- `tasks/get`: Fetch the current state of a Task by ID.\n- `tasks/cancel`: Request cancellation (currently not supported; returns an error).\n\n**LangGraph Mapping:**\n- `message.contextId` maps to LangGraph `thread_id`\n- `message.taskId` maps to LangGraph `run_id` (for resuming interrupted tasks)\n\n**Notes:**\n- Only `text` and `data` parts are supported; `file` parts are not.\n- If `message.contextId` is omitted, a new context is created.\n- Text parts require the assistant input schema to include a `messages` field.\n",
        "operationId": "post_a2a",
        "parameters": [
          {
            "name": "assistant_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "The ID of the assistant to communicate with"
          },
          {
            "name": "Accept",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "For `message/stream`, must include `text/event-stream`. For all other methods, use `application/json`."
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "jsonrpc": {
                    "type": "string",
                    "enum": [
                      "2.0"
                    ],
                    "description": "JSON-RPC version"
                  },
                  "id": {
                    "type": "string",
                    "description": "Request identifier"
                  },
                  "method": {
                    "type": "string",
                    "enum": [
                      "message/send",
                      "message/stream",
                      "tasks/get",
                      "tasks/cancel"
                    ],
                    "description": "The method to invoke"
                  },
                  "params": {
                    "type": "object",
                    "description": "Method parameters; shape depends on the method."
                  }
                },
                "required": [
                  "jsonrpc",
                  "id",
                  "method"
                ]
              },
              "examples": {
                "message_send": {
                  "summary": "Send a message (synchronous)",
                  "value": {
                    "jsonrpc": "2.0",
                    "id": "1",
                    "method": "message/send",
                    "params": {
                      "message": {
                        "role": "user",
                        "parts": [
                          {
                            "kind": "text",
                            "text": "Hello from A2A"
                          }
                        ],
                        "messageId": "msg-1",
                        "contextId": "f5bd2a40-74b6-4f7a-b649-ea3f09890003"
                      }
                    }
                  }
                },
                "message_stream": {
                  "summary": "Send a message (streaming)",
                  "value": {
                    "jsonrpc": "2.0",
                    "id": "2",
                    "method": "message/stream",
                    "params": {
                      "message": {
                        "role": "user",
                        "parts": [
                          {
                            "kind": "text",
                            "text": "Stream this response"
                          }
                        ],
                        "messageId": "msg-2"
                      }
                    }
                  }
                },
                "tasks_get": {
                  "summary": "Get a task",
                  "value": {
                    "jsonrpc": "2.0",
                    "id": "3",
                    "method": "tasks/get",
                    "params": {
                      "id": "thread-uuid:run-uuid",
                      "contextId": "thread-uuid",
                      "historyLength": 5
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "JSON-RPC response for non-streaming methods. For `message/stream`, the response is an SSE stream of JSON-RPC envelopes.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "jsonrpc": {
                      "type": "string",
                      "enum": [
                        "2.0"
                      ]
                    },
                    "id": {
                      "type": "string"
                    },
                    "result": {
                      "type": "object",
                      "description": "Success result containing task information"
                    },
                    "error": {
                      "type": "object",
                      "properties": {
                        "code": {
                          "type": "integer"
                        },
                        "message": {
                          "type": "string"
                        }
                      },
                      "description": "Error information if request failed"
                    }
                  },
                  "required": [
                    "jsonrpc",
                    "id"
                  ]
                },
                "example": {
                  "jsonrpc": "2.0",
                  "id": "1",
                  "result": {
                    "kind": "task",
                    "id": "thread-uuid:run-uuid",
                    "contextId": "thread-uuid",
                    "status": {
                      "state": "completed"
                    },
                    "artifacts": [
                      {
                        "artifactId": "artifact-uuid",
                        "name": "Assistant Response",
                        "parts": [
                          {
                            "kind": "text",
                            "text": "Hello back!"
                          }
                        ]
                      }
                    ]
                  }
                }
              },
              "text/event-stream": {
                "schema": {
                  "type": "string",
                  "description": "SSE stream of JSON-RPC response objects."
                }
              }
            }
          },
          "400": {
            "description": "Bad Request (invalid JSON-RPC, invalid params, or missing Accept header)"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Assistant not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    }
  },
...
}